// packages/bespoke/lib/game/CDASurvey/interface.d.ts
export interface ICreateParams {
}
// ...
export interface IGameState {

    /** GameState gameStage */
    gameStage?: (number|null);
}
// ...
export interface IPlayerState {

    /** PlayerState seatNumber */
    seatNumber?: (number|null);

    /** PlayerState surveyStage */
    surveyStage?: (number|null);

    /** PlayerState surveyBasic */
    surveyBasic?: (string[]|null);

    /** PlayerState surveyFeedback */
    surveyFeedback?: (string[]|null);

    /** PlayerState surveyTest */
    surveyTest?: (string[]|null);
}
// ...
export interface IMoveParams {

    /** MoveParams seatNumber */
    seatNumber?: (number|null);

    /** MoveParams inputs */
    inputs?: (string[]|null);

    /** MoveParams stage */
    stage?: (number|null);
}
// ...
export interface IPushParams {
}